refactor(server): rename gateway/gateway to gateway/worker-dispatch
The directory repeated its parent's name, so the path segment carried no
information, and its index.ts held class WorkerGateway (1,256 lines) — the
object the whole dispatch path turns on, unfindable by filename.

gateway/gateway/          -> gateway/worker-dispatch/
gateway/gateway/index.ts  -> gateway/worker-dispatch/worker-gateway.ts

Pure rename. No logic, symbol, or signature changes; the class is not
split. Comment references to the old path updated in the same pass.

(cherry picked from commit 0ed414ab2668dbff99e817329bf040b61967101e)
