+++ packages/server/src/gateway/routes/public/agent-config.ts
@@ -29,7 +29,7 @@ import {
 } from "../../auth/settings/resolved-settings-view.js";
 import type { SettingsTokenPayload } from "../../auth/settings/token-service.js";
 import type { UserAgentsStore } from "../../auth/user-agents-store.js";
-import type { WorkerConnectionManager } from "../../gateway/connection-manager.js";
+import type { WorkerConnectionManager } from "../../worker-dispatch/connection-manager.js";
 import type { IMessageQueue } from "../../infrastructure/queue/index.js";
 import type { GrantStore } from "../../permissions/grant-store.js";
 import { createOwnershipResolver } from "../shared/agent-ownership.js";
